TP4/requirements.txt
cycler==0.11.0
Cython==0.29.28
fonttools==4.31.2
joblib==1.1.0
kiwisolver==1.4.0
matplotlib==3.5.1
numpy==1.21.5
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
pyDML==0.1.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
threadpoolctl==3.1.0
typing_extensions==4.1.1
